topic/string/28. Implement strStr()_test.go
package string

import (
	"reflect"
	"testing"
)

func Test_getNext(t *testing.T) {
	type args struct {
		needle string
	}
	tests := []struct {
		name string
		args args
		want []int
	}{
		{
			name: "one",
			args: args{
				needle: "aabaaf",
			},
			want: []int{0, 1, 0, 1, 2, 0},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := getNext(tt.args.needle); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("getNext() = %v, want %v", got, tt.want)
			}
		})
	}
}

func Test_strStr(t *testing.T) {
	type args struct {
		haystack string
		needle   string
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{
			name: "one",
			args: args{
				haystack: "aabaabaafa",
				needle:   "aabaaf",
			},
			want: 3,
		},
		{
			name: "two",
			args: args{
				haystack: "hello",
				needle:   "ll",
			},
			want: 2,
		},
		{
			name: "nil",
			args: args{
				haystack: "abc",
				needle:   "",
			},
			want: 0,
		},
		{
			name: "err",
			args: args{
				haystack: "ababaabbbbababbaabaaabaabbaaaabbabaabbbbbbabbaabbabbbabbbbbaaabaababbbaabbbabbbaabbbbaaabbababbabbbabaaabbaabbabababbbaaaaaaababbabaababaabbbbaaabbbabb",
				needle:   "abbabbbabaa",
			},
			want: 92,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := strStr(tt.args.haystack, tt.args.needle); got != tt.want {
				t.Errorf("strStr() = %v, want %v", got, tt.want)
			}
		})
	}
}
